further modularizing tests
